A lot
Spd is no longer a stat. I wanted spd to have a function where it alters the amount of attacks a player can preform. After finding several balance issues, I wanted to replace SPD with RES (resistance) to magical type attacks. Furthermore, I have added sprites for both the enemies and the player (player having a placeholder pokemon spritesheet fir now). The attack function also works for both enemies and players, and next time I will add a function for defending, and making a turn system work where the player moves are hidden when the enemy preforms an attack. Then, the turn counter will be made and an enemy will spawn every 4th turn or so. And then I will add the ability to target specific enemies with attacks.